backend/requirements.txt
fastapi>=0.110
uvicorn>=0.24
websockets>=11.0
PyJWT>=2.8
pymysql>=1.1
python-dotenv>=1.0
